- Switched all the XxxTimeSeconds variables in the emulated sound component to double precision.  I'm not 100% sure that this is the real cause of the distortion issue or that it's a working fix.  If it is, though, then the proper solution will probably be to replace the timeSeconds stuff with simple down-counters (like gbsound.txt describes). refs #39 refs #45

+++ Source/sound/sound.cpp
@@ -174,7 +174,7 @@ void Sound::Run(int ticks)
 				float actualFrequency = (float)m_machine->GetTicksPerSecond() / (float)((2048 - m_sound1Frequency) << 5);
 				float actualAmplitude = (float)m_envelope1Value / (float)0x0f;
 
-				float waveX = m_fractionalSeconds * actualFrequency;
+				double waveX = (float)m_fractionalSeconds * actualFrequency;
 				waveX -= (int)waveX;
 
 				float fSample = 1.f * actualAmplitude;
@@ -212,7 +212,7 @@ void Sound::Run(int ticks)
 				float actualFrequency = (float)m_machine->GetTicksPerSecond() / (float)((2048 - m_sound2Frequency) << 5);
 				float actualAmplitude = (float)m_envelope2Value / (float)0x0f;
 
-				float waveX = m_fractionalSeconds * actualFrequency;
+				double waveX = m_fractionalSeconds * actualFrequency;
 				waveX -= (int)waveX;
 
 				float fSample = 1.f * actualAmplitude;
@@ -238,7 +238,7 @@ void Sound::Run(int ticks)
 			{
 				float actualFrequency = (float)m_machine->GetTicksPerSecond() / (float)((2048 - m_sound3Frequency) << 5);
 
-				float waveX = m_fractionalSeconds * actualFrequency;
+				double waveX = m_fractionalSeconds * actualFrequency;
 				waveX -= (int)waveX;
 
 				int sampleIndex = (int)(waveX * 32.f);
